feat: 수량(quantity)에 long 대신 int type 사용
- MenuProduct.quantity
- OrderLineItem.quantity

+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -15,19 +15,19 @@ public class MenuProduct {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "product_id")
     private Product product;
-    private long quantity;
+    private int quantity;
 
     public MenuProduct() {
     }
 
-    public MenuProduct(Long seq, Menu menu, Product product, long quantity) {
+    public MenuProduct(Long seq, Menu menu, Product product, int quantity) {
         this.seq = seq;
         this.menu = menu;
         this.product = product;
         this.quantity = quantity;
     }
 
-    public MenuProduct(Menu menu, Product product, long quantity) {
+    public MenuProduct(Menu menu, Product product, int quantity) {
         this.menu = menu;
         this.product = product;
         this.quantity = quantity;
@@ -57,11 +57,11 @@ public class MenuProduct {
         this.product = product;
     }
 
-    public long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(final long quantity) {
+    public void setQuantity(final int quantity) {
         this.quantity = quantity;
     }
 }
