+++ src/test/java/kitchenpos/factory/OrderLineItemFixtureFactory.java
@@ -5,11 +5,11 @@ import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 
 public class OrderLineItemFixtureFactory {
-    public static OrderLineItem createOrderLineItem(Long seq, Order order, Menu menu, long quantity) {
+    public static OrderLineItem createOrderLineItem(Long seq, Order order, Menu menu, int quantity) {
         return new OrderLineItem(seq, order, menu, quantity);
     }
 
-    public static OrderLineItem createOrderLineItem(Order order, Menu menu, long quantity) {
+    public static OrderLineItem createOrderLineItem(Order order, Menu menu, int quantity) {
         return new OrderLineItem(order, menu, quantity);
     }
 }
